requirements.txt
# The Price is Bot Project Requirements
# Game-specific requirements

# ===== Core Library Dependency =====
# Install elastic-grocery-core for data generation and agent tools
# For development: pip install -e ../elastic-grocery-core
# For production: pip install git+https://github.com/elastic/elastic-grocery-core.git@main
# elastic-grocery-core @ git+https://github.com/elastic/elastic-grocery-core.git@main

# ===== Core Web Framework =====
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
starlette>=0.38.0

# ===== Data Validation & Settings =====
pydantic>=2.9.0
pydantic-core>=2.23.0
pydantic-settings>=2.1.0

# ===== HTTP Clients =====
aiohttp>=3.10.0
httpx>=0.27.0
httpcore>=1.0.0

# ===== Elasticsearch =====
elasticsearch>=8.15.0
elastic-transport>=8.15.0

# ===== AWS & LLM Services =====
boto3>=1.34.0
botocore>=1.34.0
openai>=1.51.0

# ===== Authentication & Security =====
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.9.0
cryptography>=43.0.0

# ===== Data Processing =====
pandas>=2.1.0
numpy>=1.24.0

# ===== Web & Socket Support =====
python-multipart>=0.0.12
python-socketio>=5.11.0
python-engineio>=4.10.0
websockets>=13.1.0

# ===== Observability =====
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation>=0.48b0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-asgi>=0.48b0

# ===== Utilities =====
python-dotenv>=1.0.0
requests>=2.32.0
PyYAML>=6.0.0
click>=8.1.0

# ===== Development & Testing =====
pytest>=8.3.0
pytest-asyncio>=0.21.0
pytest-anyio>=0.0.0

# ===== Supporting Libraries =====
attrs>=24.2.0
certifi>=2024.8.30
charset-normalizer>=3.4.0
idna>=3.10.0
urllib3>=2.2.0
typing_extensions>=4.12.0
